I24 SSX fixed target fixes (#79)
* Cast number of exposures as an int

Reason being that the input might be a str. (As things stand, it usually is)

* Fix scan positions in fastchip.

This is a workaround for the scanspec issue: it will write both the passed start and end value and rebin in the middle depending on the number of images, ignoring the axis increment. TO avoid this, set the end position of the scan as end-increment